+++ internal/downloader/testdata/archive.tar.gz
