+++ internal/downloader/testdata/auth_server.stripped
